+++ APP/FHD.h
@@ -53,7 +53,7 @@ typedef double				fp64;				/* double precision floating point variable (64bits) 
 
 uint16 mb_swap(uint16 usData);
 uint32 mb_swap_32(uint8 *usData);
-uint8 FHD_MakeFrame(uint8 ctrl, uint16 reg_addr, uint8 *pdata, uint16 len, uint8 *buf);
+uint16 FHD_MakeFrame(uint8 ctrl, uint16 reg_addr, uint8 *pdata, uint16 len, uint8 *buf);
 uint8 FHD_CheckFrame(uint8 *buf, uint16 len);
 uint8 FHD_GetDevAddr(void);
 uint8 *FHD_GetUserData(uint8 *buf);
